Created product grid layout prototype and navbar
Used handlebars to generate a mobile-responsive grid to display
products. For the sake of being able to visualize, I created a temporary
js file that loads in a static json string with dummy data and calls
handlebars to format it. Once we get the real data in through ajax, we
can get rid of the tempItemSamples.js file, and the accompanying line
in index.js that calls it:   tempItemSamples.tempItems()

I also added a nav bar, which we can fill in later with content.