Added placeholders for updating the cart.
This is still in progress. Added functions in events.js for adding,
updating, and removing cart items. Used "let" instead of "const" for
the cartItems variable because we're going to be updating it. Getting
linter errors for now because we're not updating the value of cartItems,
but they'll go away once we add the function to update cart values.

Added a single update user api function because whether we update, add,
or remove, it's all going to the same controller method.